Define redemption-backstop config shapes once, in zod
The redemption-backstop config shape was written twice: as hand-maintained
TypeScript interfaces in configs/shared.ts and again as zod schemas in
configs/schema.ts, with a comment on each side telling the next editor to
keep them aligned. Any field added to one and forgotten in the other would
either escape validation or fail to compile at a config site.

Keep the zod schemas as the single definition and derive the TypeScript
types from them with z.infer: RedemptionCostTerms, RedemptionCapacityModel,
RedemptionCostModel, RedemptionV9RouteReviewTerms,
RedemptionV9ComposedDexExit, and RedemptionBackstopConfig now exist once
each. The explicit `z.ZodType<...>` annotations that used to pin the schemas
to the interfaces are gone with them; the doc comments those interfaces
carried move onto the corresponding schema fields.

configs/shared.ts re-exports every one of those names so the ~30 existing
import sites are untouched, and the re-exports are `export type` only, so
zod does not become reachable at runtime from src/ through this module.

The settlement-conservatism helpers move to a new configs/settlement.ts.
schema.ts needs isRedemptionSettlementAtLeastAsConservative for its
superRefine, and leaving it in shared.ts would have made the two modules
import each other once shared.ts started sourcing its types from schema.ts.
shared.ts re-exports both helpers, so callers are unaffected.
